Kill sass compiler when LiveSync stops
In a long living process (like in NativeScript Sidekick), the LiveSync process for a project may be stopped, but this will not stop the `node-sass` running in the background.
Use the after-watch hook, which is raised when LiveSync operation stops and kill currently running node-sass compiler. This way we'll not keep running watchers in the background.